cli: move backend runner and IPC into start module
