+++ packages/cli/src/modules/start/commands/package/start/startBackend.ts
@@ -17,7 +17,7 @@
 import fs from 'fs-extra';
 import { resolve as resolvePath } from 'path';
 import { paths } from '../../../../../lib/paths';
-import { runBackend } from '../../../../../lib/runner';
+import { runBackend } from '../../../lib/runner';
 
 interface StartBackendOptions {
   targetDir: string;
